requirements.txt

notebook
pandas 
numpy
tensorflow
keras
nltk
from_root
python-box
ensure
kaggle
scikit-learn
torch
transformers
h5py
-e .